비인증회원 단계 삭제 회원가입시 즉석에서 이메일로 받은 인증번호를 입력해야 가입가능하게 변경 마이페이지에 회원정보수정 기능 추가
