+++ src/main/java/com/btl/findjob/service/UserService.java
@@ -1,5 +1,7 @@
 package com.btl.findjob.service;
 
+import java.util.List;
+
 import org.apache.ibatis.annotations.Param;
 
 import com.btl.findjob.model.UserDTO;
@@ -8,7 +10,7 @@ public interface UserService {
 
 	 public int emailchk(String user_email);
 
-	public void join_insert(@Param("user_email") String user_email,@Param("user_password") String user_password,@Param("user_name") String user_name,@Param("key") String key,@Param("authorization") int authorization,@Param("salt") String salt);
+	public void join_insert(@Param("user_email") String user_email,@Param("user_password") String user_password,@Param("user_name") String user_name,@Param("authorization") int authorization,@Param("salt") String salt);
 
 	public void snsjoin_insert(@Param("user_email") String user_email,@Param("user_name") String user_name,@Param("user_password") String user_password,@Param("authorization") int authorization,@Param("sns_key") String sns_key,@Param("sns_type") String sns_type);
 		
@@ -33,4 +35,10 @@ public interface UserService {
 	public void pw_modify(@Param("user_password") String user_password,@Param("salt") String salt,@Param("user_email") String user_email);
 
 	public String getname(String user_email);
+	
+	public List<UserDTO> user_info(String user_email);
+	
+	public void user_info_modify(@Param("user_email") String user_email,@Param("user_password") String user_password,@Param("user_name") String user_name,@Param("salt") String salt);
+
+
 }
