+++ src/main/webapp/WEB-INF/views/includes/myPage_Menu.jsp
@@ -16,7 +16,7 @@
         <div class="card-group">
             <div class="card-body">
                 <SPAN class="">
-                 ${name}님 안녕하세요.<br> 
+                 ${name}님의 마이페이지<br> 
                 <a href="user_info">회원정보 수정</a> 
                 </SPAN>
             </div>
