--- src/main/webapp/WEB-INF/views/user/auth.jsp
@@ -1,9 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-         pageEncoding="UTF-8" %>
- 
-<script>
-	alert('${auth}');
-	window.close();
-</script>
-
-
